requirements.txt
click==8.1.7
Flask==2.3.3
Jinja2==3.1.2
python-dotenv==1.0.0
Werkzeug==2.3.7